Fix publication on the 4.0.0 branch
When we use 'io.micronaut.library', we end up with multiple POMs in the published POM we generate.

We were using this to get nativeTest to pass.  Switching this to 'io.micronaut.graalvm' still allows this.

But it does not generate an invalid POM

Also updated the shared setting plugin whilst I was here