Add dynamic registration of test property providers
This commit introduces a new way to supply test properties to
tests, by implementing a `TestPropertyProviderFactory`. These
factories are loaded via service loading and are responsible
for instantiating a `TestPropertyProvider`.

In order to be able to dynamically provide properties, the
factory is provided with the current test class, and the
set of properties which are already available at the moment
the provider is called.

This means that contrary to a typical `TestPropertyProvider`,
this provider has access to "raw" configuration, as supplied
by `PropertySourceLoader`.

This will permit the implementation of a new test resources
module which will be capable of dynamically supplying test
properties to test classes _before_ the application context
is available, and without having to explicitly call the
test resources client.