Rebuild initial ctx when first test method has props
Otherwise we want to keep the original context as other extensions or
tests may expect to have the original context. This doesn't fix the
issue where these two features don't play well together, but it
drastically reduces the scope where they conflict. This means that we
should be able to access the same context over the test lifecycle unless
we are using property annotations on the test method.